+++ tests/adaptive_map/test_adaptive_map_iterator.c
@@ -211,11 +211,11 @@ check_static_begin(adaptive_map_test_iterate_removal) {
     );
     /* Seed the test with any integer for reproducible random test sequence
        currently this will change every test. NOLINTNEXTLINE */
-    srand(time(NULL));
+    srand((unsigned)time(NULL));
     size_t const num_nodes = 100;
     for (size_t i = 0; i < num_nodes; ++i) {
-        /* Force duplicates. */
-        int const key = rand() % (num_nodes + 1); // NOLINT
+        /* NOLINTNEXTLINE(cert-msc30-c, cert-msc50-cpp). */
+        int const key = (int)((size_t)rand() % (num_nodes + 1));
         (void)insert_or_assign(
             &s,
             &(struct Val){
@@ -249,11 +249,11 @@ check_static_begin(adaptive_map_test_iterate_remove_key_value_reinsert) {
     );
     /* Seed the test with any integer for reproducible random test sequence
        currently this will change every test. NOLINTNEXTLINE */
-    srand(time(NULL));
+    srand((unsigned)time(NULL));
     size_t const num_nodes = 100;
     for (size_t i = 0; i < num_nodes; ++i) {
-        /* Force duplicates. */
-        int const key = rand() % (num_nodes + 1); // NOLINT
+        /* NOLINTNEXTLINE(cert-msc30-c, cert-msc50-cpp). */
+        int const key = (int)((size_t)rand() % (num_nodes + 1));
         (void)insert_or_assign(
             &s,
             &(struct Val){
