+++ tests/array_adaptive_map/test_array_adaptive_map_insert.c
@@ -351,7 +351,7 @@ check_static_begin(array_adaptive_map_test_insert_via_array_macros) {
             &array_adaptive_map,
             insert_handle(
                 array_adaptive_map_handle_wrap(&array_adaptive_map, &i),
-                &(struct Val){i, i},
+                &(struct Val){(int)i, (int)i},
                 &(CCC_Allocator){}
             )
         );
@@ -366,7 +366,7 @@ check_static_begin(array_adaptive_map_test_insert_via_array_macros) {
             &array_adaptive_map,
             insert_handle(
                 array_adaptive_map_handle_wrap(&array_adaptive_map, &i),
-                &(struct Val){i, i + 1},
+                &(struct Val){(int)i, (int)i + 1},
                 &(CCC_Allocator){}
             )
         );
@@ -477,7 +477,7 @@ check_static_begin(array_adaptive_map_test_two_sum) {
         }
         CCC_Handle const e = insert_or_assign(
             &array_adaptive_map,
-            &(struct Val){.id = addends[i], .val = i},
+            &(struct Val){.id = addends[i], .val = (int)i},
             &(CCC_Allocator){}
         );
         check(insert_error(&e), false);
@@ -951,12 +951,12 @@ check_static_begin(array_adaptive_map_test_insert_weak_srand) {
         (CCC_Key_comparator){.compare = id_order},
         (struct Val[STANDARD_FIXED_CAP]){}
     );
-    srand(time(NULL)); /* NOLINT */
+    srand((unsigned)time(NULL)); /* NOLINT */
     for (int i = 0; i < num_nodes; ++i) {
         CCC_Handle const e = swap_handle(
             &array_adaptive_map,
             &(struct Val){
-                .id = rand() /* NOLINT */,
+                .id = (int)rand() /* NOLINT */,
                 .val = i,
             },
             &(CCC_Allocator){}
