+++ tests/array_tree_map/array_tree_map_utility.c
@@ -22,7 +22,7 @@ check_begin(
     int const larger_prime,
     CCC_Allocator const *const allocator
 ) {
-    size_t shuffled_index = larger_prime % size;
+    size_t shuffled_index = (size_t)larger_prime % size;
     for (size_t i = 0; i < size; ++i) {
         (void)insert_or_assign(
             m,
@@ -30,7 +30,7 @@ check_begin(
             allocator
         );
         check(validate(m), true);
-        shuffled_index = (shuffled_index + larger_prime) % size;
+        shuffled_index = (shuffled_index + (size_t)larger_prime) % size;
     }
     check(count(m).count, size);
     check_end();
